+++ CHANGELOG.md
@@ -13,9 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
  - **Breaking**: `Array` `retrieve_` methods now return `Vec<u8>`/`Vec<T>` instead of `Box<[u8]>`/`Box<[T]>`
    - This avoids potential internal reallocations
- - **Breaking**: Remove `StorePrefixError::new`, deprecated since `v0.7.3`
  - **Breaking**: `StoreKey::parent` now returns `StorePrefix` instead of `Option<StorePrefix>`
  - **Breaking**: `ZipStorageAdapter::{new,new_with_path}` now take a `StoreKey`
+
+### Removed
+ - **Breaking**: Remove `StorePrefixError::new`, deprecated since `v0.7.3`
+ - **Breaking**: Remove `ArraySubset::{in_subset,in_subset_unchecked}`, deprecated since `v0.7.2`
 
 ### Fixed
  - Disallow an empty string for a `StoreKey`